fix: add handle to util class
