+++ app/src/main/java/com/example/newsapp/data/repository/NewsRepository.kt
@@ -3,27 +3,13 @@ package com.example.newsapp.data.repository
 import com.example.newsapp.data.api.ApiService
 import com.example.newsapp.data.api.NetworkResult
 import com.example.newsapp.data.entity.NewsDto
+import com.example.newsapp.data.repository.util.NetworkUtil
 import retrofit2.Response
 import javax.inject.Inject
 
 class NewsRepository @Inject constructor(
     private val apiService: ApiService
 ) {
-    suspend fun<T: Any> handleApi(
-        execute: suspend () -> Response<T>
-    ) : NetworkResult<T> {
-        return try {
-            val response = execute()
-            val body = response.body()
-            if (response.isSuccessful && body != null) {
-                NetworkResult.Success(body)
-            } else {
-                NetworkResult.Error(response.code(), response.message())
-            }
-        } catch (e: Exception) {
-            NetworkResult.Error(e.hashCode(), e.message)
-        }
-    }
-
-    suspend fun dataSource() :NetworkResult<NewsDto> = handleApi { apiService.getNews() }
+    suspend fun dataSource(): NetworkResult<NewsDto> =
+        NetworkUtil().handleApi { apiService.getNews() }
 }
