+++ app/src/main/java/com/example/newsapp/data/repository/util/NetworkUtil.kt
@@ -1,0 +1,22 @@
+package com.example.newsapp.data.repository.util
+
+import com.example.newsapp.data.api.NetworkResult
+import retrofit2.Response
+
+class NetworkUtil {
+    suspend fun <T : Any> handleApi(
+        execute: suspend () -> Response<T>
+    ): NetworkResult<T> {
+        return try {
+            val response = execute()
+            val body = response.body()
+            if (response.isSuccessful && body != null) {
+                NetworkResult.Success(body)
+            } else {
+                NetworkResult.Error(response.code(), response.message())
+            }
+        } catch (e: Exception) {
+            NetworkResult.Error(e.hashCode(), e.message)
+        }
+    }
+}
